Separate history pruning function and history pruning store files
