+++ endpointmanager/pkg/historypruning/historypruning_integration_test.go
@@ -1,6 +1,6 @@
 // +build integration
 
-package postgresql
+package historypruning
 
 import (
 	"context"
